+++ generator/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 1.0.0-beta.2
+
+* Update for modified entitiy visitor
+* Update for entity and feature finders
+
 ## 1.0.0-beta.1
 
 * Initial release